Migrate escalation requests out of admin namespace
