+++ app/views/escalation_requests/index.haml
@@ -4,10 +4,10 @@
       View Escalation Requests
   .container-fluid
     .row-fluid
-      = link_to "All", admin_escalation_requests_path(status: ""), class: "btn btn-default"
-      = link_to "Pending", admin_escalation_requests_path(status: "pending"), class: "btn btn-primary"
-      = link_to "Approved", admin_escalation_requests_path(status: "approved"), class: "btn btn-success"
-      = link_to "Denied", admin_escalation_requests_path(status: "denied"), class: "btn btn-danger"
+      = link_to "All", escalation_requests_path(status: ""), class: "btn btn-default"
+      = link_to "Pending", escalation_requests_path(status: "pending"), class: "btn btn-primary"
+      = link_to "Approved", escalation_requests_path(status: "approved"), class: "btn btn-success"
+      = link_to "Denied", escalation_requests_path(status: "denied"), class: "btn btn-danger"
       .table-responsive
         %table.table
           %thead
@@ -43,11 +43,11 @@
                 %td
                   = request.escalation_end_at
                 %td
-                  %a.btn.btn-xs.btn-success{href: admin_escalation_request_approve_path(request), data: {method: 'post'}}
+                  %a.btn.btn-xs.btn-success{href: escalation_request_approve_path(request), data: {method: 'post'}}
                     %i.fa.fa-check
-                  %a.btn.btn-xs.btn-danger{href: admin_escalation_request_deny_path(request), data: {method: 'post'}}
+                  %a.btn.btn-xs.btn-danger{href: escalation_request_deny_path(request), data: {method: 'post'}}
                     %i.fa.fa-times
-                  %a.btn.btn-xs.btn-primary{href: edit_admin_escalation_request_path(request, "#{request.escalatable.class.to_s.downcase}_id" => request.escalatable.id )}
+                  %a.btn.btn-xs.btn-primary{href: edit_escalation_request_path(request, "#{request.escalatable.class.to_s.downcase}_id" => request.escalatable.id )}
                     %i.fa.fa-gear
               %tr
                 %td.hiddenRow{colspan: "6"}
